[Refactor] #102 - Usecase package
