+++ domain/src/main/java/com/woowahan/domain/usecase/banchan/FetchSoupDishBanchanUseCase.kt
@@ -1,16 +1,17 @@
-package com.woowahan.domain.usecase
+package com.woowahan.domain.usecase.banchan
 
 import com.woowahan.domain.model.BanchanModel
 import com.woowahan.domain.repository.BanchanRepository
+import com.woowahan.domain.usecase.cart.FetchCartItemsKeyUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
-class FetchSideDishBanchanUseCase(
+class FetchSoupDishBanchanUseCase(
     private val banchanRepository: BanchanRepository,
     private val fetchCartItemsKeyUseCase: FetchCartItemsKeyUseCase
 ) {
     suspend operator fun invoke(): Flow<Result<List<BanchanModel>>> {
-        return banchanRepository.fetchSideDishBanchan().map {
+        return banchanRepository.fetchSoupDishBanchan().map {
             kotlin.runCatching {
                 val cart = fetchCartItemsKeyUseCase().getOrThrow()
                 listOf(
