+++ domain/src/main/java/com/woowahan/domain/usecase/cart/InsertCartItemUseCase.kt
@@ -1,4 +1,4 @@
-package com.woowahan.domain.usecase
+package com.woowahan.domain.usecase.cart
 
 import com.woowahan.domain.model.BanchanModel
 import com.woowahan.domain.repository.CartRepository
