+++ presentation/src/main/java/com/woowahan/banchan/ui/viewmodel/RootViewModel.kt
@@ -2,7 +2,7 @@ package com.woowahan.banchan.ui.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.woowahan.domain.usecase.GetCartItemsSizeFlowUseCase
+import com.woowahan.domain.usecase.cart.GetCartItemsSizeFlowUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
